refactor: refactored chat domain
