+++ src/main/java/teamproject/aipro/domain/chat/controller/ChatController.java
@@ -1,22 +1,18 @@
 package teamproject.aipro.domain.chat.controller;
 
-import org.springframework.beans.factory.annotation.Value;
+import java.security.Principal;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
 import teamproject.aipro.domain.chat.dto.request.ChatRequest;
 import teamproject.aipro.domain.chat.dto.response.ChatResponse;
-import teamproject.aipro.domain.chat.entity.ChatCatalog;
-import teamproject.aipro.domain.chat.service.ChatHistoryService;
 import teamproject.aipro.domain.chat.service.ChatService;
 
 @RestController
@@ -24,70 +20,29 @@ import teamproject.aipro.domain.chat.service.ChatService;
 @RequestMapping("/api/chat")
 public class ChatController {
 
-	@Value("${jwt.secret}")
-	private String secretKey;
-
 	private final ChatService chatService;
-	private final ChatHistoryService chatHistoryService;
 
-	public ChatController(ChatService chatService, ChatHistoryService chatHistoryService) {
+	public ChatController(ChatService chatService) {
 		this.chatService = chatService;
-		this.chatHistoryService = chatHistoryService;
 	}
 
 	@PostMapping("/question")
 	public ResponseEntity<ChatResponse> question(
-			@RequestHeader("Authorization") String authHeader,
+			Principal principal,
 			@RequestBody ChatRequest chatRequest,
 			@RequestParam(required = false) String catalogId) {
 
-		// Authorization 헤더에서 토큰을 파싱하여 사용자 ID 추출
-		String userId = extractUserIdFromToken(authHeader);
+		String userId = principal.getName();
 		System.out.println("userId = " + userId);
+
 		if (userId == null) {
 			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
 		}
 
 		ChatResponse response = (catalogId == null || catalogId.trim().isEmpty())
-				? processNewCatalogRequest(chatRequest, userId)
-				: processExistingCatalogRequest(chatRequest, catalogId, userId);
+				? chatService.processNewCatalogRequest(chatRequest, userId)
+				: chatService.processExistingCatalogRequest(chatRequest, catalogId, userId);
 
 		return ResponseEntity.ok(response);
-	}
-
-	// 토큰에서 사용자 ID 추출 메서드
-	private String extractUserIdFromToken(String authHeader) {
-		try {
-			if (authHeader == null) {
-				return null; // 토큰이 없거나 잘못된 형식인 경우
-			}
-			String jwtToken = authHeader.replace("Bearer ", "");
-			Claims claims = Jwts.parserBuilder()
-					.setSigningKey(secretKey.getBytes()) // secretKey를 JWT 서명 검증에 사용
-					.build()
-					.parseClaimsJws(jwtToken)
-					.getBody();
-			return claims.getSubject(); // JWT의 subject 필드에서 userId 반환
-		} catch (Exception e) {
-			System.err.println("Error decoding JWT: " + e.getMessage());
-			return null; // 토큰 검증 실패 시 null 반환
-		}
-	}
-
-	private ChatResponse processNewCatalogRequest(ChatRequest chatRequest, String userId) {
-		// AI 서버로부터 요약 받기
-		ChatResponse response = chatHistoryService.summary(chatRequest);
-		Long newCatalogId = createNewCatalog(userId, response.getMessage());
-		// 새로운 ChatHistory 저장
-		return chatService.question(chatRequest, String.valueOf(newCatalogId), userId);
-	}
-
-	private Long createNewCatalog(String userId, String summaryMessage) {
-		ChatCatalog chatCatalog = new ChatCatalog(userId, summaryMessage);
-		return chatHistoryService.saveChatCatalog(chatCatalog.getUserId(), chatCatalog.getChatSummary()).getId();
-	}
-
-	private ChatResponse processExistingCatalogRequest(ChatRequest chatRequest, String catalogId, String userId) {
-		return chatService.question(chatRequest, catalogId, userId);
 	}
 }
