+++ src/main/java/teamproject/aipro/domain/chat/service/ChatService.java
@@ -10,6 +10,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import teamproject.aipro.domain.chat.dto.request.AiRequest;
 import teamproject.aipro.domain.chat.dto.request.ChatRequest;
 import teamproject.aipro.domain.chat.dto.response.ChatResponse;
+import teamproject.aipro.domain.chat.entity.ChatCatalog;
 import teamproject.aipro.domain.role.service.RoleService;
 
 @Service
@@ -51,5 +52,22 @@ public class ChatService {
 			System.err.println("Error occurred while calling AI server: " + e.getMessage());
 			return new ChatResponse("Error: Unable to get response from AI server.", catalogId);
 		}
+	}
+
+	public ChatResponse processNewCatalogRequest(ChatRequest chatRequest, String userId) {
+		// AI 서버로부터 요약 받기
+		ChatResponse response = chatHistoryService.summary(chatRequest);
+		Long newCatalogId = createNewCatalog(userId, response.getMessage());
+		// 새로운 ChatHistory 저장
+		return question(chatRequest, String.valueOf(newCatalogId), userId);
+	}
+
+	public ChatResponse processExistingCatalogRequest(ChatRequest chatRequest, String catalogId, String userId) {
+		return question(chatRequest, catalogId, userId);
+	}
+
+	private Long createNewCatalog(String userId, String summaryMessage) {
+		ChatCatalog chatCatalog = new ChatCatalog(userId, summaryMessage);
+		return chatHistoryService.saveChatCatalog(chatCatalog.getUserId(), chatCatalog.getChatSummary()).getId();
 	}
 }
